refactor(auth): 인증 로직에서 사용되었던 필드 제거에 따른 클래스 위치 이동
